had to add  constexpr positiveElseZero to prevent negative values to kill the template recursion of nop.
couldnt get the loop to properly NOT unroll, so i just upped the value that is needed to use the part with the loop. i didnt even count cycles to see if it works out, but at least it uses the loop for now. have to further investigate this another time

delayNanoseconds should work for now